zios/Zeus/requirements.txt
# System Requirements
# ------------------
# - Python 3.8+
# - Golang (go) 1.16+
# - GCC (C Compiler)
# - libcurl-dev
# - libjson-c-dev

# Python Packages
# --------------
requests>=2.28.0
dnspython>=2.2.0
python-whois>=0.7.3
pyOpenSSL>=22.0.0
Pillow>=9.0.0
numpy>=1.21.0
cryptography>=37.0.0
beautifulsoup4>=4.9.3
colorama>=0.4.4
urllib3>=1.26.0
pycurl>=7.45.0
json-c>=0.15.0

# Note: Before installing Python packages, ensure you have:
# For Ubuntu/Debian:
#   sudo apt-get install golang gcc libcurl4-openssl-dev libjson-c-dev
#
# For CentOS/RHEL:
#   sudo yum install golang gcc libcurl-devel json-c-devel
#
# For Windows:
#   1. Install Go from https://golang.org/dl/
#   2. Install MinGW (GCC) from http://mingw-w64.org/
#   3. Add both to system PATH
